+++ TabbedPanelMod/history.md
@@ -1,5 +1,10 @@
 # History
 
+## v0.3
+
+* Added init script to change Tabbed Panel layout on Freplane's start up.
+* Added runOnStartingFreeplane preference
+
 ## v0.2
 
 * Added screen shot image